FossilOrigin-Name: 8c0ede76112819e18cc4fc26e75a3ead8f6309eec61d70dd51ec53b8362bff03

+++ src/btree.c
@@ -5120,7 +5120,7 @@ static int accessPayload(
 
   getCellInfo(pCur);
   aPayload = pCur->info.pPayload;
-  assert( offset+amt <= pCur->info.nPayload );
+  assert( (u64)offset+(u64)amt <= (u64)pCur->info.nPayload );
 
   assert( aPayload > pPage->aData );
   if( (uptr)(aPayload - pPage->aData) > (pBt->usableSize - pCur->info.nLocal) ){
